Library Reference

CategoricalTensorNetworks.TensorNetworkAlgebras — Module

Tensor networks as algebras of the operad of undirected wiring diagrams.

CategoricalTensorNetworks.TensorNetworkAlgebras.contract_tensor_network — Method

Contract a tensor network all at once.

Performs a generalized contraction of a list of tensors of arbitrary rank. In the binary case, this function includes matrix multiplication, tensor product, and trace as special cases. In general, it implements the "generalized matrix multiplication" described in Spivak et al's "Pixel Arrays".

Be warned that this method is inefficient for contracting networks with more than two junctions. Specifically, if the network has $n$ junctions with dimensions $k₁, …, kₙ$, then this function performs exactly $k₁ ⋯ kₙ$ multiplications in the base ring (or rig). For example, a binary matrix multiply of square matrices performs $k³$ multiplications. A ternary matrix multiply performs $k⁴$ multiplications, whereas a sequence of two binary multiplies performs $2k³$. Thus, when there are more than two junctions, it is usually better to schedule a computation using schedule or another package, possibly via @contract_tensors_with.

In addition to the method taking an undirected wiring diagram, this function has methods similar to the ncon ("Network CONtractor") function in:

the NCON package for MATLAB

the Julia package TensorOperations.jl

except that here the outer junctions are represented explictly by a third argument rather than implicitly by using negative numbers in the second argument.

CategoricalTensorNetworks.TensorNetworkAlgebras.gen_tensor_notation — Method

Generate Julia expression in tensor notation from undirected wiring diagram.

This function is used to implement @contract_tensors_with but may be useful in its own right.

See also @tensor_network, the "inverse" to this function.

CategoricalTensorNetworks.TensorNetworkAlgebras.parse_tensor_network — Method

Parse an undirected wiring diagram from a tensor expression.

For more information, see the corresponding macro @tensor_network.

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Method

Contract tensors according to UWD.

This method simply wraps contract_tensor_network, which also presents an interface closer to the standard tensor algebra packages.

CategoricalTensorNetworks.TensorNetworkAlgebras.@contract_tensors_with — Macro

Contract a tensor network using a macro from another package.

This macro takes two arguments: an undirected wiring diagram and a macro call supporting the tensor contraction notation that is de facto standard among Julia tensor packages. For example, to evaluate a tensor network using the @tullio macro, use:

A = @contract_tensors_with @tullio diagram

The following macros should work:

@tensor from TensorOperations.jl.

@tullio from Tullio.jl

@einsum from Einsum.jl

@ein from OMEinsum.jl

However, the macros @cast and @reduce from TensorCast.jl will not work because they do not support implicit summation.

CategoricalTensorNetworks.TensorNetworkAlgebras.@tensor_network — Macro

Construct an undirected wiring diagram using tensor notation.

The syntax is compatible with the Einstein-style tensor notation used by tensor algebra packages like TensorOperations.jl and TensorCast.jl. For example, the wiring diagram for the composite of two matrices (or two binary relations) is constructed by

@tensor_network (i,j,k) C[i,k] := A[i,j] * B[j,k]

The leading context, or list of variables, may be omitted, in which case it is inferred from the variables used in the tensor expression. So, in this example, an equivalent macro call is

@tensor_network C[i,k] := A[i,j] * B[j,k]

See also: gen_tensor_notation, the "inverse" to this macro.

CategoricalTensorNetworks.ScheduleUWDs — Module

Scheduling and evaluation of undirected wiring diagrams.

In category-theoretic terms, this module is about evaluating arbitrary composites of morphisms in hypergraph categories.

CategoricalTensorNetworks.ScheduleUWDs.HasNestedUWD — Type

Abstract type for C-sets that contained a nested UWD.

CategoricalTensorNetworks.ScheduleUWDs.HasScheduledUWD — Type

Abstract type for C-sets that contain a scheduled UWDS.

This type includes ScheduledUWD and NestedUWD.

CategoricalTensorNetworks.ScheduleUWDs.NestedUWD — Type

Nested undirected wiring diagram.

A nested UWD is a scheduled UWD whose composite nodes have been given ports, making explicit the intermediate boxes in the composition.

Nested UWDs are very similar but not quite identical to Robin Milner's bigraphs.

See also: ScheduledUWD.

CategoricalTensorNetworks.ScheduleUWDs.ScheduledUWD — Type

Scheduled undirected wiring diagram.

A scheduled UWD consists of a UWD together with a set of composite nodes forming a rooted tree, or in general a rooted forest, whose leaves are the diagram's boxes.

See also: NestedUWD.

CategoricalTensorNetworks.ScheduleUWDs.SequentialSchedule — Type

Schedule diagram as a sequential chain of binary composites.

This is the simplest possible scheduling algorithm, the equivalent of foldl for undirected wiring diagrams. Unless otherwise specified, the boxes are folded according to the order of their IDs, which is arbitrary.

Base.schedule — Method

Schedule an undirected wiring diagram.

By default, a simple sequential schedule is used.

CategoricalTensorNetworks.ScheduleUWDs.eval_schedule — Method

Evaluate a scheduled UWD given a set of generators for the boxes.

The optional first argument f should be a callable with the same signature as oapply for UWD algebras, which by default is oapply itself.

f(composite::UndirectedWiringDiagram, values::AbstractVector)

Nested UWDs are used as an auxiliary data structure during the evaluation. If the schedule will be evaluated multiple times, you should explicitly convert it to a nested UWD using to_nested_diagram and then call eval_schedule on the resulting object instead.

CategoricalTensorNetworks.ScheduleUWDs.to_nested_diagram — Method

Convert a scheduled UWD to a nested UWD.